+++ packages/chains/src/base.ts
@@ -1,0 +1,36 @@
+import { Chain } from './types'
+
+export const base = {
+  id: 8453,
+  network: 'base',
+  name: 'Base',
+  nativeCurrency: { name: 'Base', symbol: 'ETH', decimals: 18 },
+  rpcUrls: {
+    default: {
+      http: ['https://mainnet.base.org'],
+    },
+    public: {
+      http: ['https://mainnet.base.org'],
+    },
+  },
+  blockExplorers: {
+    blockscout: {
+      name: 'Basescout',
+      url: 'https://base.blockscout.com',
+    },
+    default: {
+      name: 'Basescan',
+      url: 'https://basescan.org',
+    },
+    etherscan: {
+      name: 'Basescan',
+      url: 'https://basescan.org',
+    },
+  },
+  contracts: {
+    multicall3: {
+      address: '0xca11bde05977b3631167028862be2a173976ca11',
+      blockCreated: 5022,
+    },
+  },
+} as const satisfies Chain