+++ packages/chains/src/gnosis.ts
@@ -16,11 +16,17 @@ export const gnosis = {
   blockExplorers: {
     etherscan: {
       name: 'Gnosisscan',
-      url: 'https://gnosisscan.io/',
+      url: 'https://gnosisscan.io',
     },
     default: {
       name: 'Gnosis Chain Explorer',
-      url: 'https://blockscout.com/xdai/mainnet/',
+      url: 'https://blockscout.com/xdai/mainnet',
+    },
+  },
+  contracts: {
+    multicall3: {
+      address: '0xca11bde05977b3631167028862be2a173976ca11',
+      blockCreated: 21022491,
     },
   },
 } as const satisfies Chain