+++ packages/chains/src/skale/chaosTestnet.ts
@@ -10,15 +10,21 @@ export const skaleChaosTestnet = {
       http: [
         'https://staging-v3.skalenodes.com/v1/staging-fast-active-bellatrix',
       ],
+      webSocket: [
+        'wss://staging-v3.skalenodes.com/v1/ws/staging-fast-active-bellatrix',
+      ],
     },
     public: {
       http: [
         'https://staging-v3.skalenodes.com/v1/staging-fast-active-bellatrix',
       ],
+      webSocket: [
+        'wss://staging-v3.skalenodes.com/v1/ws/staging-fast-active-bellatrix',
+      ],
     },
   },
   blockExplorers: {
-    etherscan: {
+    blockscout: {
       name: 'SKALE Explorer',
       url: 'https://staging-fast-active-bellatrix.explorer.staging-v3.skalenodes.com',
     },
@@ -27,6 +33,11 @@ export const skaleChaosTestnet = {
       url: 'https://staging-fast-active-bellatrix.explorer.staging-v3.skalenodes.com',
     },
   },
-  contracts: {},
+  contracts: {
+    multicall3: {
+      address: '0xcA11bde05977b3631167028862bE2a173976CA11',
+      blockCreated: 1192202,
+    },
+  },
   testnet: true,
 } as const satisfies Chain