+++ packages/chains/src/skale/cryptoBlades.ts
@@ -8,13 +8,19 @@ export const skaleCryptoBlades = {
   rpcUrls: {
     default: {
       http: ['https://mainnet.skalenodes.com/v1/affectionate-immediate-pollux'],
+      webSocket: [
+        'wss://mainnet.skalenodes.com/v1/ws/affectionate-immediate-pollux',
+      ],
     },
     public: {
       http: ['https://mainnet.skalenodes.com/v1/affectionate-immediate-pollux'],
+      webSocket: [
+        'wss://mainnet.skalenodes.com/v1/ws/affectionate-immediate-pollux',
+      ],
     },
   },
   blockExplorers: {
-    etherscan: {
+    blockscout: {
       name: 'SKALE Explorer',
       url: 'https://affectionate-immediate-pollux.explorer.mainnet.skalenodes.com',
     },