+++ packages/chains/src/skale/cryptoColosseum.ts
@@ -8,13 +8,15 @@ export const skaleCryptoColosseum = {
   rpcUrls: {
     default: {
       http: ['https://mainnet.skalenodes.com/v1/haunting-devoted-deneb'],
+      webSocket: ['wss://mainnet.skalenodes.com/v1/ws/haunting-devoted-deneb'],
     },
     public: {
       http: ['https://mainnet.skalenodes.com/v1/haunting-devoted-deneb'],
+      webSocket: ['wss://mainnet.skalenodes.com/v1/ws/haunting-devoted-deneb'],
     },
   },
   blockExplorers: {
-    etherscan: {
+    blockscout: {
       name: 'SKALE Explorer',
       url: 'https://haunting-devoted-deneb.explorer.mainnet.skalenodes.com',
     },